SocialService-DefineDomainModel: Refactored the packaging name, and the project name in pom.xml
